+++ src/app-shell-parts/useAppShellSections.ts
@@ -6,37 +6,10 @@ import {
   useRef,
   useState,
 } from "react";
-import { homeDir } from "@tauri-apps/api/path";
 import { revealItemInDir } from "@tauri-apps/plugin-opener";
-import {
-  captureBrowserAgentSnapshot,
-  ensureWorkspacePathDir,
-  getWorkspaceFiles,
-  isWebServiceRuntime,
-} from "../services/tauri";
+import { getWorkspaceFiles } from "../services/tauri";
 import { pushErrorToast } from "../services/toasts";
 import {
-  isKanbanThreadCompatibleWithEngine,
-  resolveKanbanThreadCreationStrategy,
-} from "../features/kanban/utils/contextMode";
-import { deriveKanbanTaskTitle } from "../features/kanban/utils/taskTitle";
-import { findTaskDownstream } from "../features/kanban/utils/chaining";
-import { buildChainedPromptPrefix, extractKanbanResultSnapshot } from "../features/kanban/utils/resultSnapshot";
-import {
-  beginKanbanTaskRunLifecycle,
-  patchKanbanTaskRunLifecycle,
-} from "../features/tasks/utils/kanbanTaskRunLifecycle";
-import {
-  beginTaskRunRecovery,
-  cancelTaskRunRecovery,
-} from "../features/tasks/utils/taskRunRecovery";
-import { buildLatestRunSummary } from "../features/tasks/utils/taskRunProjection";
-import {
-  deriveTaskRunTelemetryPatch,
-} from "../features/tasks/utils/taskRunTelemetry";
-import {
-  buildTaskRunBrowserEvidenceRef,
-  loadTaskRunStore,
   patchTaskRun,
   saveTaskRunStore,
 } from "../features/tasks/utils/taskRunStorage";
@@ -47,24 +20,6 @@ import {
   saveOrchestrationTaskStore,
   upsertOrchestrationTask,
 } from "../features/agent-orchestration";
-import {
-  buildBrowserContextAttachment,
-  getActiveBrowserContext,
-} from "../features/browser-agent";
-import type { TaskRunRecord } from "../features/tasks/types";
-import {
-  applyMissedRunPolicy,
-  hasReachedRecurringRoundLimit,
-  isScheduleDue,
-  markRecurringScheduleCompleted,
-  markScheduleTriggered,
-  resolvePostProcessingStatus,
-} from "../features/kanban/utils/scheduling";
-import type {
-  KanbanTask,
-  KanbanTaskExecutionSource,
-  KanbanTaskStatus,
-} from "../features/kanban/types";
 import { useSoloMode } from "../features/layout/hooks/useSoloMode";
 import { useLiveEditPreview } from "../features/live-edit-preview/hooks/useLiveEditPreview";
 import { useArchiveShortcut } from "../features/app/hooks/useArchiveShortcut";
@@ -75,7 +30,6 @@ import { useAppMenuEvents } from "../features/app/hooks/useAppMenuEvents";
 import { useMenuAcceleratorController } from "../features/app/hooks/useMenuAcceleratorController";
 import { useMenuLocalization } from "../features/app/hooks/useMenuLocalization";
 import { runWithLoadingProgress } from "../features/app/utils/loadingProgressActions";
-import { isDefaultWorkspacePath } from "../features/workspaces/utils/defaultWorkspace";
 import { normalizeSharedSessionEngine } from "../features/shared-session/utils/sharedSessionEngines";
 import {
   buildDetachedSpecHubSession,
@@ -83,13 +37,11 @@ import {
 } from "../features/spec/detachedSpecHub";
 import { openOrFocusClientDocumentationWindow } from "../features/client-documentation/clientDocumentationWindow";
 import type { WorkspaceHomeDeleteResult } from "../features/workspaces/components/WorkspaceHome";
-import type { EngineType, MessageSendOptions, WorkspaceInfo } from "../types";
-import type { KanbanContextMode } from "../features/kanban/utils/contextMode";
+import type { EngineType, WorkspaceInfo } from "../types";
 import {
   isRewindSupportedThreadId,
   resolvePendingSessionThreadCandidate,
   resolveTaskThreadId,
-  stripComposerKanbanTagsPreserveFormatting,
   syncKanbanExecutionEngineAndModel,
 } from "./useAppShellSections.kanbanHelpers";
 import {
@@ -97,6 +49,8 @@ import {
   shouldCollapseRightPanelOnThreadSelect,
   shouldPreserveEditorOnThreadSelect,
 } from "./threadEditorPreservation";
+import { useAppShellKanbanComposerSection } from "./useAppShellKanbanComposerSection";
+import { useAppShellKanbanExecutionSection } from "./useAppShellKanbanExecutionSection";
 export {
   resolvePendingSessionThreadCandidate,
   resolveTaskThreadId,
@@ -104,9 +58,6 @@ export {
   stripComposerKanbanTagsPreserveFormatting,
   syncKanbanExecutionEngineAndModel,
 } from "./useAppShellSections.kanbanHelpers";
-
-const KANBAN_SCHEDULER_INTERVAL_MS = 20_000;
-const KANBAN_EXECUTION_LOCK_STALE_MS = 120_000;
 
 export function useAppShellSections(ctx: any) {
   const {
@@ -236,409 +187,16 @@ export function useAppShellSections(ctx: any) {
     resolveCanonicalThreadId,
   } = ctx;
 
-  const [selectedComposerKanbanPanelId, setSelectedComposerKanbanPanelId] =
-    useState<string | null>(null);
-  const [composerKanbanContextMode, setComposerKanbanContextMode] =
-    useState<KanbanContextMode>("new");
-  const composerKanbanWorkspacePaths = useMemo(() => {
-    if (!activeWorkspace) {
-      return [] as string[];
-    }
-    const paths = new Set<string>();
-    paths.add(activeWorkspace.path);
-    if (activeWorkspace.parentId) {
-      const parentWorkspace = workspaces.find((workspace) => workspace.id === activeWorkspace.parentId);
-      if (parentWorkspace) {
-        paths.add(parentWorkspace.path);
-      }
-    }
-    // If current workspace is a parent/main workspace, include its worktrees too.
-    for (const workspace of workspaces) {
-      if (workspace.parentId === activeWorkspace.id) {
-        paths.add(workspace.path);
-      }
-    }
-    return Array.from(paths);
-  }, [activeWorkspace, workspaces]);
-  const composerLinkedKanbanPanels = useMemo(() => {
-    if (composerKanbanWorkspacePaths.length === 0) {
-      return [];
-    }
-    return kanbanPanels
-      .filter((panel) => composerKanbanWorkspacePaths.includes(panel.workspaceId))
-      .slice()
-      .sort((a, b) => b.createdAt - a.createdAt || a.sortOrder - b.sortOrder)
-      .map((panel) => ({
-        id: panel.id,
-        name: panel.name,
-        workspaceId: panel.workspaceId,
-        createdAt: panel.createdAt,
-      }));
-  }, [composerKanbanWorkspacePaths, kanbanPanels]);
-
-  useEffect(() => {
-    if (!selectedComposerKanbanPanelId) {
-      return;
-    }
-    const stillExists = composerLinkedKanbanPanels.some(
-      (panel) => panel.id === selectedComposerKanbanPanelId,
-    );
-    if (!stillExists) {
-      setSelectedComposerKanbanPanelId(null);
-    }
-  }, [composerLinkedKanbanPanels, selectedComposerKanbanPanelId]);
-
-  const handleOpenComposerKanbanPanel = useCallback(
-    (panelId: string) => {
-      const panel = composerLinkedKanbanPanels.find((entry) => entry.id === panelId);
-      if (!panel) {
-        return;
-      }
-      setKanbanViewState({
-        view: "board",
-        workspaceId: panel.workspaceId,
-        panelId,
-      });
-      setAppMode("kanban");
-    },
-    [composerLinkedKanbanPanels, setAppMode, setKanbanViewState],
-  );
-
-  const resolveComposerKanbanPanel = useCallback(
-    (text: string) => {
-      const tagMatches = Array.from(text.matchAll(/&@([^\s]+)/g))
-        .map((entry) => entry[1]?.trim())
-        .filter((value): value is string => Boolean(value));
-      const panelByName = new Map(
-        composerLinkedKanbanPanels.map((panel) => [panel.name, panel.id]),
-      );
-      const firstTaggedPanelId =
-        tagMatches.map((name) => panelByName.get(name)).find(Boolean) ?? null;
-      const panelId =
-        firstTaggedPanelId ??
-        (selectedComposerKanbanPanelId &&
-        composerLinkedKanbanPanels.some(
-          (panel) => panel.id === selectedComposerKanbanPanelId,
-        )
-          ? selectedComposerKanbanPanelId
-          : null);
-      const cleanText = stripComposerKanbanTagsPreserveFormatting(text);
-      return { panelId, cleanText };
-    },
-    [composerLinkedKanbanPanels, selectedComposerKanbanPanelId],
-  );
-
-  const mergeSelectedAgentOption = useCallback(
-    (options?: MessageSendOptions): MessageSendOptions | undefined => {
-      if (activeEngine === "opencode") {
-        return options;
-      }
-      const selectedAgentForSend =
-        selectedAgentRef?.current ?? selectedAgent ?? null;
-      const merged: MessageSendOptions = {
-        ...(options ?? {}),
-        selectedAgent: selectedAgentForSend
-          ? {
-              id: selectedAgentForSend.id,
-              name: selectedAgentForSend.name,
-              prompt: selectedAgentForSend.prompt ?? null,
-              icon: selectedAgentForSend.icon ?? null,
-            }
-          : null,
-      };
-      return merged;
-    },
-    [activeEngine, selectedAgent, selectedAgentRef],
-  );
-
-  const handleComposerSendWithKanban = useCallback(
-    async (
-      text: string,
-      images: string[],
-      options?: MessageSendOptions,
-    ) => {
-      const trimmedOriginalText = text.trim();
-      const { panelId, cleanText } = resolveComposerKanbanPanel(trimmedOriginalText);
-      const textForSending = cleanText;
-
-      // HomeChat send: no active workspace yet. Select or create one, then
-      // create a thread and jump to normal chat view before sending.
-      if (!activeWorkspaceId && !isPullRequestComposer) {
-        let workspace: WorkspaceInfo | null = null;
-        if (isWebServiceRuntime()) {
-          workspace =
-            workspaces.find((entry) => isDefaultWorkspacePath(entry.path)) ??
-            workspaces.find((entry) => entry.kind === "main") ??
-            workspaces[0] ??
-            null;
-
-          if (!workspace) {
-            try {
-              const resolvedHome = normalizePath(await homeDir());
-              if (!resolvedHome) {
-                throw new Error("Unable to resolve default workspace path.");
-              }
-              const preferredPaths = [
-                `${resolvedHome}/.ccgui/workspace`,
-                `${resolvedHome}/.mossx/workspace`,
-                `${resolvedHome}/.codemoss/workspace`,
-              ];
-
-              let createdWorkspacePath: string | null = null;
-              let lastError: unknown = null;
-              for (const candidatePath of preferredPaths) {
-                try {
-                  await ensureWorkspacePathDir(candidatePath);
-                  createdWorkspacePath = candidatePath;
-                  break;
-                } catch (error) {
-                  lastError = error;
-                }
-              }
-              if (!createdWorkspacePath) {
-                throw lastError ?? new Error("Failed to create default workspace path.");
-              }
-              const normalizedDefaultPath = normalizePath(createdWorkspacePath);
-              workspace = workspaces.find(
-                (entry) => normalizePath(entry.path) === normalizedDefaultPath,
-              ) ?? null;
-              if (!workspace) {
-                workspace = await addWorkspaceFromPath(createdWorkspacePath);
-              }
-            } catch (error) {
-              alertError(error);
-              return;
-            }
-          }
-        } else {
-          let defaultWorkspacePath: string;
-          try {
-            const resolvedHome = normalizePath(await homeDir());
-            defaultWorkspacePath = `${resolvedHome}/.ccgui/workspace`;
-            await ensureWorkspacePathDir(defaultWorkspacePath);
-          } catch (error) {
-            alertError(error);
-            return;
-          }
-          const normalizedDefaultPath = normalizePath(defaultWorkspacePath);
-          workspace = workspaces.find(
-            (entry) => normalizePath(entry.path) === normalizedDefaultPath,
-          ) ?? null;
-          if (!workspace) {
-            try {
-              workspace = await addWorkspaceFromPath(defaultWorkspacePath);
-            } catch (error) {
-              alertError(error);
-              return;
-            }
-          }
-        }
-
-        if (!workspace) {
-          return;
-        }
-        exitDiffView();
-        resetPullRequestSelection();
-        setWorkspaceHomeWorkspaceId(null);
-        setAppMode("chat");
-        setCenterMode("chat");
-        selectWorkspace(workspace.id);
-        if (!workspace.connected) {
-          await connectWorkspace(workspace);
-        }
-        const threadId = await startThreadForWorkspace(workspace.id, {
-          engine: activeEngine,
-          activate: true,
-        });
-        if (!threadId) {
-          return;
-        }
-        setActiveThreadId(threadId, workspace.id);
-        const fallbackText =
-          textForSending.length > 0 ? textForSending : trimmedOriginalText;
-        if (fallbackText.length > 0 || images.length > 0) {
-          await sendUserMessageToThread(
-            workspace,
-            threadId,
-            fallbackText,
-            images,
-            mergeSelectedAgentOption(options),
-          );
-        }
-        return;
-      }
-
-      if (!panelId || !activeWorkspaceId || isPullRequestComposer) {
-        const fallbackText =
-          textForSending.length > 0 ? textForSending : trimmedOriginalText;
-        await handleComposerSend(
-          fallbackText,
-          images,
-          mergeSelectedAgentOption(options),
-        );
-        return;
-      }
-
-      const workspace = workspacesById.get(activeWorkspaceId);
-      if (!workspace) {
-        await handleComposerSend(
-          textForSending.length > 0 ? textForSending : trimmedOriginalText,
-          images,
-          mergeSelectedAgentOption(options),
-        );
-        return;
-      }
-
-      // &@ 看板消息必须在新会话里执行，不能污染当前会话窗口
-      if (!workspace.connected) {
-        await connectWorkspace(workspace);
-      }
-      const engine = (activeEngine === "codex" ? "codex" : "claude") as
-        | "codex"
-        | "claude";
-      const activeThreadEngine =
-        activeThreadId && activeWorkspaceId
-          ? (
-              threadsByWorkspace[activeWorkspaceId]?.find(
-                (thread) => thread.id === activeThreadId,
-              )?.engineSource ?? null
-            )
-          : null;
-      const isActiveThreadInWorkspace = Boolean(
-        activeWorkspaceId &&
-          activeThreadId &&
-          threadsByWorkspace[activeWorkspaceId]?.some(
-            (thread) => thread.id === activeThreadId,
-          ),
-      );
-      const threadCreationStrategy = resolveKanbanThreadCreationStrategy({
-        mode: composerKanbanContextMode,
-        engine,
-        activeThreadId,
-        activeThreadEngine,
-        activeWorkspaceId,
-        targetWorkspaceId: workspace.id,
-        isActiveThreadInWorkspace,
-      });
-      const canInheritViaFork = threadCreationStrategy === "inherit";
-      const threadId =
-        canInheritViaFork && activeThreadId
-          ? await forkThreadForWorkspace(activeWorkspaceId, activeThreadId, {
-              activate: false,
-            })
-          : await startThreadForWorkspace(activeWorkspaceId, {
-              engine,
-              activate: false,
-            });
-      const resolvedThreadId =
-        threadId ??
-        (await startThreadForWorkspace(activeWorkspaceId, {
-          engine,
-          activate: false,
-        }));
-      if (!resolvedThreadId) {
-        return;
-      }
-      if (canInheritViaFork && !threadId) {
-        addDebugEntry({
-          id: `${Date.now()}-kanban-linked-fork-fallback`,
-          timestamp: Date.now(),
-          source: "client",
-          label: "kanban/linked fork fallback",
-          payload: {
-            workspaceId: activeWorkspaceId,
-            reason: "fork-unavailable",
-          },
-        });
-      }
-
-      if (textForSending.length > 0 || images.length > 0) {
-        await sendUserMessageToThread(
-          workspace,
-          resolvedThreadId,
-          textForSending,
-          images,
-          mergeSelectedAgentOption(options),
-        );
-      }
-
-      const taskDescription = textForSending.length > 0 ? textForSending : trimmedOriginalText;
-      const taskFallbackTitle =
-        composerLinkedKanbanPanels.find((panel) => panel.id === panelId)?.name ||
-        "Kanban Task";
-      const taskTitle = deriveKanbanTaskTitle(taskDescription, taskFallbackTitle);
-      const createdTask = kanbanCreateTask({
-        workspaceId: workspace.path,
-        panelId,
-        title: taskTitle,
-        description: taskDescription,
-        engineType: engine,
-        modelId: effectiveSelectedModelId,
-        branchName: "main",
-        images,
-        autoStart: true,
-      });
-
-      kanbanUpdateTask(createdTask.id, {
-        threadId: resolvedThreadId,
-        status: "inprogress",
-      });
-    },
-    [
-      resolveComposerKanbanPanel,
-      handleComposerSend,
-      mergeSelectedAgentOption,
-      activeWorkspaceId,
-      normalizePath,
-      addWorkspaceFromPath,
-      alertError,
-      workspaces,
-      workspacesById,
-      exitDiffView,
-      resetPullRequestSelection,
-      selectWorkspace,
-      setAppMode,
-      setActiveThreadId,
-      setCenterMode,
-      setWorkspaceHomeWorkspaceId,
-      connectWorkspace,
-      startThreadForWorkspace,
-      forkThreadForWorkspace,
-      sendUserMessageToThread,
-      isPullRequestComposer,
-      activeEngine,
-      activeThreadId,
-      threadsByWorkspace,
-      addDebugEntry,
-      composerKanbanContextMode,
-      effectiveSelectedModelId,
-      composerLinkedKanbanPanels,
-      kanbanCreateTask,
-      kanbanUpdateTask,
-    ],
-  );
-
-  const handleComposerSendWithEditorFallback = useCallback(
-    async (
-      text: string,
-      images: string[],
-      options?: MessageSendOptions,
-    ) => {
-      await handleComposerSendWithKanban(text, images, options);
-    },
-    [handleComposerSendWithKanban],
-  );
-
-  const handleComposerQueueWithEditorFallback = useCallback(
-    async (
-      text: string,
-      images: string[],
-      options?: MessageSendOptions,
-    ) => {
-      await handleComposerQueue(text, images, mergeSelectedAgentOption(options));
-    },
-    [handleComposerQueue, mergeSelectedAgentOption],
-  );
+  const {
+    selectedComposerKanbanPanelId,
+    setSelectedComposerKanbanPanelId,
+    composerKanbanContextMode,
+    setComposerKanbanContextMode,
+    composerLinkedKanbanPanels,
+    handleOpenComposerKanbanPanel,
+    handleComposerSendWithEditorFallback,
+    handleComposerQueueWithEditorFallback,
+  } = useAppShellKanbanComposerSection(ctx);
 
   const handleRewindFromMessage = useCallback(
     async (
@@ -990,1304 +548,17 @@ export function useAppShellSections(ctx: any) {
     [clearDraftForThread, removeImagesForThread, removeThread, removeThreads, t],
   );
 
-  const kanbanTasksRef = useRef(kanbanTasks);
-  const schedulerStartedAtRef = useRef(Date.now());
-  const kanbanExecutionLocksRef = useRef<
-    Record<string, { token: string; source: KanbanTaskExecutionSource; acquiredAt: number }>
-  >({});
-
-  useEffect(() => {
-    kanbanTasksRef.current = kanbanTasks;
-  }, [kanbanTasks]);
-
-  const updateTaskExecution = useCallback(
-    (taskId: string, changes: Record<string, unknown>) => {
-      const current = kanbanTasksRef.current.find((task) => task.id === taskId);
-      if (!current) {
-        return;
-      }
-      kanbanUpdateTask(taskId, {
-        execution: {
-          ...(current.execution ?? {}),
-          ...changes,
-        },
-      });
-    },
-    [kanbanUpdateTask],
-  );
-
-  const setTaskChainBlockedReason = useCallback(
-    (taskId: string, blockedReason: string | null) => {
-      const current = kanbanTasksRef.current.find((task) => task.id === taskId);
-      if (!current?.chain) {
-        return;
-      }
-      kanbanUpdateTask(taskId, {
-        chain: {
-          ...current.chain,
-          blockedReason,
-        },
-      });
-    },
-    [kanbanUpdateTask],
-  );
-
-  const patchTaskRunAndProjectToKanban = useCallback(
-    (input: Parameters<typeof patchKanbanTaskRunLifecycle>[0]) => {
-      let result: ReturnType<typeof patchKanbanTaskRunLifecycle> = null;
-      try {
-        result = patchKanbanTaskRunLifecycle(input);
-      } catch (error) {
-        console.error("Failed to patch Kanban task run lifecycle", error);
-        return null;
-      }
-      if (result?.latestRunSummary) {
-        kanbanUpdateTask(result.run.task.taskId, {
-          latestRunSummary: result.latestRunSummary,
-        });
-      }
-      return result;
-    },
-    [kanbanUpdateTask],
-  );
-
-  const persistKanbanTaskComposerSelection = useCallback(
-    (workspaceId: string, threadId: string, modelId: string | null | undefined) => {
-      if (!modelId) {
-        return;
-      }
-      const currentSelection = resolveComposerSelectionForThread(workspaceId, threadId);
-      persistComposerSelectionForThread(workspaceId, threadId, {
-        modelId,
-        effort: currentSelection?.effort ?? null,
-      });
-    },
-    [persistComposerSelectionForThread, resolveComposerSelectionForThread],
-  );
-
-  const handleDispatchOrchestrationTask = useCallback(
-    async (confirmation: any): Promise<{ ok: boolean; taskId?: string | null; reason?: string }> => {
-      const taskId = confirmation?.task?.taskId ?? null;
-      const validEngines = new Set(["codex", "claude", "gemini"]);
-      const validThreadStrategies = new Set(["new_thread", "reuse_active_thread", "choose_thread"]);
-      if (
-        !confirmation?.task ||
-        typeof confirmation.workspaceId !== "string" ||
-        !validEngines.has(confirmation.engine) ||
-        !validThreadStrategies.has(confirmation.threadStrategy)
-      ) {
-        return { ok: false, taskId, reason: "invalid_dispatch_confirmation" };
-      }
-
-      const initial = beginOrchestrationTaskDispatch({
-        ...confirmation,
-        persist: false,
-      });
-      if (!initial.ok) {
-        return { ok: false, taskId, reason: initial.reason };
-      }
-      saveTaskRunStore(initial.taskRunStore);
-      saveOrchestrationTaskStore(initial.orchestrationTaskStore);
-
-      const startedAt = Date.now();
-      let threadId: string | null = null;
-      try {
-        const workspace =
-          activeWorkspace?.id === confirmation.workspaceId
-            ? activeWorkspace
-            : workspaces.find((entry: WorkspaceInfo) => entry.id === confirmation.workspaceId) ?? null;
-        if (!workspace) {
-          throw new Error("workspace_not_found");
-        }
-
-        await connectWorkspace(workspace);
-        await setActiveEngine(confirmation.engine);
-        threadId =
-          confirmation.threadStrategy === "reuse_active_thread"
-            ? confirmation.task.linkedSessionIds[0] ?? null
-            : null;
-        if (!threadId) {
-          threadId = await startThreadForWorkspace(workspace.id, {
-            engine: confirmation.engine,
-            activate: true,
-          });
-        }
-        if (!threadId) {
-          throw new Error("thread_create_failed");
-        }
-
-        if (confirmation.model) {
-          const currentSelection = resolveComposerSelectionForThread(workspace.id, threadId);
-          persistComposerSelectionForThread(workspace.id, threadId, {
-            modelId: confirmation.model,
-            effort: currentSelection?.effort ?? null,
-          });
-        }
-
-        setActiveThreadId(threadId, workspace.id);
-        await sendUserMessageToThread(
-          workspace,
-          threadId,
-          buildOrchestrationDispatchPrompt(confirmation),
-          [],
-          confirmation.model ? { model: confirmation.model } : undefined,
-        );
-
-        const nextTaskRunStore = patchTaskRun(initial.taskRunStore, initial.run.runId, {
-          status: "running",
-          model: confirmation.model ?? null,
-          linkedThreadId: threadId,
-          currentStep: "first_message_sent",
-          latestOutputSummary: "Task prompt sent to session.",
-          startedAt,
-          now: startedAt,
-        });
-        const nextOrchestrationTaskStore = patchOrchestrationTask(
-          upsertOrchestrationTask(initial.orchestrationTaskStore, confirmation.task),
-          confirmation.task.taskId,
-          {
-            status: "running",
-            preferredEngine: confirmation.engine,
-            preferredModel: confirmation.model ?? null,
-            linkedSessionIds: [...new Set([...confirmation.task.linkedSessionIds, threadId])],
-            now: new Date(startedAt).toISOString(),
-          },
-        );
-        saveTaskRunStore(nextTaskRunStore);
-        saveOrchestrationTaskStore(nextOrchestrationTaskStore);
-        return { ok: true, taskId: confirmation.task.taskId };
-      } catch (error) {
-        const reason = error instanceof Error ? error.message : String(error);
-        const failedAt = Date.now();
-        const linkedSessionIds = threadId
-          ? [...new Set([...(confirmation.task.linkedSessionIds ?? []), threadId])]
-          : confirmation.task.linkedSessionIds ?? [];
-        const nextTaskRunStore = patchTaskRun(initial.taskRunStore, initial.run.runId, {
-          status: "failed",
-          model: confirmation.model ?? null,
-          linkedThreadId: threadId ?? initial.run.linkedThreadId ?? null,
-          currentStep: "runtime_start_failed",
-          latestOutputSummary: reason,
-          failureReason: reason,
-          availableRecoveryActions: ["open_conversation", "retry", "fork_new_run"],
-          finishedAt: failedAt,
-          now: failedAt,
-        });
-        const nextOrchestrationTaskStore = patchOrchestrationTask(
-          initial.orchestrationTaskStore,
-          confirmation.task.taskId,
-          {
-            status: "blocked",
-            preferredEngine: confirmation.engine,
-            preferredModel: confirmation.model ?? null,
-            linkedSessionIds,
-            now: new Date(failedAt).toISOString(),
-          },
-        );
-        saveTaskRunStore(nextTaskRunStore);
-        saveOrchestrationTaskStore(nextOrchestrationTaskStore);
-        return { ok: false, taskId: confirmation.task.taskId, reason };
-      }
-    },
-    [
-      activeWorkspace,
-      connectWorkspace,
-      persistComposerSelectionForThread,
-      resolveComposerSelectionForThread,
-      sendUserMessageToThread,
-      setActiveEngine,
-      setActiveThreadId,
-      startThreadForWorkspace,
-      workspaces,
-    ],
-  );
-
-  const launchKanbanTaskExecution = useCallback(
-    async (params: {
-      taskId: string;
-      source: KanbanTaskExecutionSource;
-      activate?: boolean;
-      injectedPrefix?: string;
-      forceNewThread?: boolean;
-      existingRunId?: string | null;
-    }): Promise<{ ok: true; threadId: string } | { ok: false; reason: string }> => {
-      const task = kanbanTasksRef.current.find((entry) => entry.id === params.taskId);
-      if (!task) {
-        return { ok: false, reason: "task_not_found" };
-      }
-      let taskRunId: string | null = params.existingRunId ?? null;
-      let launchedSuccessfully = false;
-      if (params.source !== "chained" && task.chain?.previousTaskId) {
-        setTaskChainBlockedReason(task.id, "chain_requires_head_trigger");
-        updateTaskExecution(task.id, {
-          lastSource: params.source,
-          blockedReason: "chain_requires_head_trigger",
-        });
-        return { ok: false, reason: "chain_requires_head_trigger" };
-      }
-      if (params.source === "chained" && task.chain?.previousTaskId) {
-        setTaskChainBlockedReason(task.id, null);
-      }
-      const existingLock = kanbanExecutionLocksRef.current[task.id];
-      if (existingLock) {
-        updateTaskExecution(task.id, {
-          lastSource: params.source,
-          blockedReason: "non_reentrant_trigger_blocked",
-        });
-        return { ok: false, reason: "non_reentrant_trigger_blocked" };
-      }
-      if (!taskRunId) {
-        let taskRunResult: ReturnType<typeof beginKanbanTaskRunLifecycle> | null = null;
-        try {
-          taskRunResult = beginKanbanTaskRunLifecycle({
-            task,
-            source: params.source,
-          });
-        } catch (error) {
-          console.error("Failed to begin Kanban task run lifecycle", error);
-        }
-        if (taskRunResult && !taskRunResult.ok) {
-          if (taskRunResult.latestRunSummary) {
-            kanbanUpdateTask(task.id, {
-              latestRunSummary: taskRunResult.latestRunSummary,
-            });
-          }
-          updateTaskExecution(task.id, {
-            lastSource: params.source,
-            blockedReason: taskRunResult.reason,
-          });
-          return { ok: false, reason: taskRunResult.reason };
-        }
-        if (taskRunResult) {
-          taskRunId = taskRunResult.run.runId;
-          kanbanUpdateTask(task.id, {
-            latestRunSummary: taskRunResult.latestRunSummary,
-          });
-        }
-      }
-
-      const lock = {
-        token: `${params.source}-${Date.now()}`,
-        source: params.source,
-        acquiredAt: Date.now(),
-      } as const;
-      kanbanExecutionLocksRef.current[task.id] = lock;
-      updateTaskExecution(task.id, {
-        lastSource: params.source,
-        lock,
-        blockedReason: null,
-      });
-
-      try {
-        const workspace = workspacesByPath.get(task.workspaceId);
-        if (!workspace) {
-          throw new Error("workspace_not_found");
-        }
-
-        await connectWorkspace(workspace);
-        const engine = (task.engineType ?? activeEngine) as "claude" | "codex";
-        const workspaceThreads = threadsByWorkspace[workspace.id] ?? [];
-        const { outboundModel, shouldSyncComposerSelection, composerSelection } =
-          await syncKanbanExecutionEngineAndModel({
-            activate: params.activate,
-            engine,
-            modelId: task.modelId,
-            setActiveEngine,
-          });
-
-        const shouldForceNewThread = Boolean(params.forceNewThread);
-        const canonicalTaskThreadId =
-          shouldForceNewThread
-            ? null
-            : resolveTaskThreadId(task.threadId, resolveCanonicalThreadId);
-        const canonicalTaskThreadEngine =
-          canonicalTaskThreadId
-            ? (
-                workspaceThreads.find((entry) => entry.id === canonicalTaskThreadId)
-                  ?.engineSource ?? null
-              )
-            : null;
-        const canReuseExistingThread = isKanbanThreadCompatibleWithEngine({
-          engine,
-          threadId: canonicalTaskThreadId,
-          threadEngine: canonicalTaskThreadEngine,
-        });
-        let threadId = canReuseExistingThread ? canonicalTaskThreadId : null;
-        if (shouldForceNewThread && task.threadId) {
-          // Keep previous run in review state before switching task to the new execution thread.
-          kanbanUpdateTask(task.id, { status: "testing" });
-        }
-        if (
-          canonicalTaskThreadId &&
-          canonicalTaskThreadId !== task.threadId &&
-          canReuseExistingThread
-        ) {
-          kanbanUpdateTask(task.id, { threadId: canonicalTaskThreadId });
-          patchTaskRunAndProjectToKanban({
-            runId: taskRunId,
-            status: "planning",
-            linkedThreadId: canonicalTaskThreadId,
-            currentStep: "reuse_existing_thread",
-          });
-        }
-        if (!threadId) {
-          threadId = await startThreadForWorkspace(workspace.id, {
-            engine,
-            activate: params.activate ?? false,
-          });
-          if (!threadId) {
-            throw new Error("thread_create_failed");
-          }
-          kanbanUpdateTask(task.id, { threadId });
-          patchTaskRunAndProjectToKanban({
-            runId: taskRunId,
-            status: "planning",
-            linkedThreadId: threadId,
-            currentStep: "thread_created",
-          });
-        }
-
-        if (shouldSyncComposerSelection && composerSelection?.modelId) {
-          persistKanbanTaskComposerSelection(workspace.id, threadId, composerSelection.modelId);
-        }
-
-        const executionStartedAt = Date.now();
-        const baseMessage = task.description?.trim() || task.title;
-        let browserContextAttachment = null;
-        try {
-          const activeBrowserContext = getActiveBrowserContext();
-          if (
-            activeBrowserContext &&
-            activeBrowserContext.workspaceId === workspace.id &&
-            activeBrowserContext.rendererBound &&
-            activeBrowserContext.session.status === "ready"
-          ) {
-            const snapshot = await captureBrowserAgentSnapshot(
-              activeBrowserContext.browserSessionId,
-            );
-            browserContextAttachment = buildBrowserContextAttachment(snapshot);
-          }
-        } catch (error) {
-          console.warn(
-            "Browser context source evidence unavailable for Kanban task dispatch",
-            error,
-          );
-        }
-        const firstMessage = params.injectedPrefix
-          ? `${params.injectedPrefix}\n\n${baseMessage}`
-          : baseMessage;
-        if (firstMessage) {
-          await sendUserMessageToThread(workspace, threadId, firstMessage, task.images ?? [], {
-            ...(outboundModel ? { model: outboundModel } : {}),
-            ...(browserContextAttachment ? { browserContextAttachment } : {}),
-          });
-        }
-
-        kanbanUpdateTask(task.id, { status: "inprogress" });
-        patchTaskRunAndProjectToKanban({
-          runId: taskRunId,
-          status: "running",
-          linkedThreadId: threadId,
-          currentStep: "first_message_sent",
-          startedAt: executionStartedAt,
-          browserEvidence: browserContextAttachment
-            ? buildTaskRunBrowserEvidenceRef(browserContextAttachment)
-            : undefined,
-        });
-        updateTaskExecution(task.id, {
-          lastSource: params.source,
-          lock: null,
-          blockedReason: null,
-          startedAt: executionStartedAt,
-          finishedAt: null,
-        });
-        launchedSuccessfully = true;
-        return { ok: true, threadId };
-      } catch (error) {
-        const reason = error instanceof Error ? error.message : String(error);
-        updateTaskExecution(task.id, {
-          lastSource: params.source,
-          lock: null,
-          blockedReason: reason,
-        });
-        patchTaskRunAndProjectToKanban({
-          runId: taskRunId,
-          status: "failed",
-          failureReason: reason,
-          finishedAt: Date.now(),
-        });
-        return { ok: false, reason };
-      } finally {
-        if (!launchedSuccessfully) {
-          delete kanbanExecutionLocksRef.current[task.id];
-        }
-      }
-    },
-    [
-      workspacesByPath,
-      connectWorkspace,
-      activeEngine,
-      threadsByWorkspace,
-      setActiveEngine,
-      startThreadForWorkspace,
-      persistKanbanTaskComposerSelection,
-      kanbanUpdateTask,
-      sendUserMessageToThread,
-      updateTaskExecution,
-      setTaskChainBlockedReason,
-      resolveCanonicalThreadId,
-      patchTaskRunAndProjectToKanban,
-    ],
-  );
-
-  // --- Kanban conversation handlers ---
-  const handleOpenTaskConversation = useCallback(
-    async (task: KanbanTask) => {
-      setSelectedKanbanTaskId(task.id);
-      const workspace = workspacesByPath.get(task.workspaceId);
-      if (!workspace) return;
-
-      await connectWorkspace(workspace);
-      selectWorkspace(workspace.id);
-
-      const engine = (task.engineType ?? activeEngine) as "claude" | "codex";
-      const workspaceThreads = threadsByWorkspace[workspace.id] ?? [];
-      await setActiveEngine(engine);
-
-      if (task.threadId) {
-        let resolvedThreadId =
-          resolveTaskThreadId(task.threadId, resolveCanonicalThreadId) ?? task.threadId;
-        const resolvedThreadEngine =
-          workspaceThreads.find((entry) => entry.id === resolvedThreadId)?.engineSource ?? null;
-        const canReuseExistingThread = isKanbanThreadCompatibleWithEngine({
-          engine,
-          threadId: resolvedThreadId,
-          threadEngine: resolvedThreadEngine,
-        });
-        if (resolvedThreadId !== task.threadId) {
-          kanbanUpdateTask(task.id, { threadId: resolvedThreadId });
-        }
-
-        if (!canReuseExistingThread) {
-          resolvedThreadId = "";
-        }
-
-        const isPendingThread =
-          resolvedThreadId.startsWith("claude-pending-") ||
-          resolvedThreadId.startsWith("opencode-pending-");
-        const hasThreadStatus = threadStatusById[resolvedThreadId] !== undefined;
-        const existsInWorkspaceThreads = workspaceThreads.some(
-          (entry) => entry.id === resolvedThreadId,
-        );
-
-        if (isPendingThread && !hasThreadStatus && !existsInWorkspaceThreads) {
-          const occupiedThreadIds = new Set(
-            kanbanTasks
-              .filter((entry) => entry.id !== task.id && entry.threadId)
-              .map((entry) =>
-                resolveTaskThreadId(entry.threadId, resolveCanonicalThreadId),
-              )
-              .filter(
-                (threadId): threadId is string =>
-                  Boolean(
-                    threadId &&
-                      !threadId.startsWith("claude-pending-") &&
-                      !threadId.startsWith("opencode-pending-"),
-                  ),
-              ),
-          );
-          const uniqueCandidate = resolvePendingSessionThreadCandidate({
-            pendingThreadId: resolvedThreadId,
-            workspaceThreadIds: workspaceThreads.map((entry) => entry.id),
-            occupiedThreadIds,
-          });
-          if (uniqueCandidate) {
-            resolvedThreadId = uniqueCandidate;
-            kanbanUpdateTask(task.id, { threadId: resolvedThreadId });
-          }
-        }
-
-        const canActivateExistingThread =
-          threadStatusById[resolvedThreadId] !== undefined ||
-          workspaceThreads.some((entry) => entry.id === resolvedThreadId) ||
-          resolvedThreadId.startsWith("claude-pending-") ||
-          resolvedThreadId.startsWith("opencode-pending-");
-        if (canActivateExistingThread) {
-          persistKanbanTaskComposerSelection(workspace.id, resolvedThreadId, task.modelId);
-          setActiveThreadId(resolvedThreadId, workspace.id);
-          return;
-        }
-      }
-
-      const threadId = await startThreadForWorkspace(workspace.id, { engine });
-      if (threadId) {
-        kanbanUpdateTask(task.id, { threadId });
-        persistKanbanTaskComposerSelection(workspace.id, threadId, task.modelId);
-        setActiveThreadId(threadId, workspace.id);
-      }
-    },
-    [
-      workspacesByPath,
-      connectWorkspace,
-      selectWorkspace,
-      setActiveThreadId,
-      startThreadForWorkspace,
-      kanbanUpdateTask,
-      activeEngine,
-      setActiveEngine,
-      persistKanbanTaskComposerSelection,
-      threadStatusById,
-      threadsByWorkspace,
-      kanbanTasks,
-      resolveCanonicalThreadId,
-      setSelectedKanbanTaskId,
-    ]
-  );
-
-  const handleCloseTaskConversation = useCallback(() => {
-    setSelectedKanbanTaskId(null);
-  }, [setSelectedKanbanTaskId]);
-
-  const resolveTaskByRun = useCallback((run: TaskRunRecord) => {
-    return kanbanTasksRef.current.find((task) => task.id === run.task.taskId) ?? null;
-  }, []);
-
-  const handleRetryTaskRun = useCallback(
-    (run: TaskRunRecord) => {
-      const task = resolveTaskByRun(run);
-      if (!task) {
-        return;
-      }
-      const recovery = beginTaskRunRecovery({
-        task,
-        trigger: "retry",
-        parentRun: run,
-      });
-      if (!recovery.ok) {
-        if (recovery.latestRunSummary) {
-          kanbanUpdateTask(task.id, { latestRunSummary: recovery.latestRunSummary });
-        }
-        return;
-      }
-      kanbanUpdateTask(task.id, { latestRunSummary: recovery.latestRunSummary });
-      void launchKanbanTaskExecution({
-        taskId: task.id,
-        source: "manual",
-        activate: false,
-        forceNewThread: true,
-        existingRunId: recovery.run.runId,
-      });
-    },
-    [kanbanUpdateTask, launchKanbanTaskExecution, resolveTaskByRun],
-  );
-
-  const handleForkTaskRun = useCallback(
-    (run: TaskRunRecord) => {
-      const task = resolveTaskByRun(run);
-      if (!task) {
-        return;
-      }
-      const recovery = beginTaskRunRecovery({
-        task,
-        trigger: "forked",
-        parentRun: run,
-      });
-      if (!recovery.ok) {
-        if (recovery.latestRunSummary) {
-          kanbanUpdateTask(task.id, { latestRunSummary: recovery.latestRunSummary });
-        }
-        return;
-      }
-      kanbanUpdateTask(task.id, { latestRunSummary: recovery.latestRunSummary });
-      void launchKanbanTaskExecution({
-        taskId: task.id,
-        source: "manual",
-        activate: false,
-        forceNewThread: true,
-        existingRunId: recovery.run.runId,
-      });
-    },
-    [kanbanUpdateTask, launchKanbanTaskExecution, resolveTaskByRun],
-  );
-
-  const handleResumeTaskRun = useCallback(
-    async (run: TaskRunRecord) => {
-      if (!run.linkedThreadId) {
-        return;
-      }
-      const task = resolveTaskByRun(run);
-      if (task) {
-        await handleOpenTaskConversation(task);
-      }
-    },
-    [handleOpenTaskConversation, resolveTaskByRun],
-  );
-
-  const handleCancelTaskRun = useCallback(
-    async (run: TaskRunRecord) => {
-      if (run.linkedThreadId && activeThreadId === run.linkedThreadId) {
-        await interruptTurn();
-      }
-      const canceled = cancelTaskRunRecovery({ runId: run.runId, now: Date.now() });
-      if (canceled.run) {
-        kanbanUpdateTask(run.task.taskId, {
-          latestRunSummary: buildLatestRunSummary(canceled.run),
-        });
-      }
-    },
-    [activeThreadId, interruptTurn, kanbanUpdateTask],
-  );
-
-  const handleKanbanCreateTask = useCallback(
-    (input: Parameters<typeof kanbanCreateTask>[0]) => {
-      const task = kanbanCreateTask(input);
-      if (input.autoStart) {
-        const tryLaunch = (attempt: number) => {
-          void launchKanbanTaskExecution({
-            taskId: task.id,
-            source: "autoStart",
-            activate: false,
-          }).then((result) => {
-            if (result.ok) {
-              return;
-            }
-            if (result.reason !== "task_not_found" || attempt >= 3) {
-              return;
-            }
-            window.setTimeout(() => {
-              tryLaunch(attempt + 1);
-            }, (attempt + 1) * 40);
-          });
-        };
-        tryLaunch(0);
-      }
-      return task;
-    },
-    [kanbanCreateTask, launchKanbanTaskExecution],
-  );
-
-  // Sync kanban task threadIds when pending IDs are renamed to session IDs.
-  // Strategy:
-  // 1) Prefer canonical alias resolution from useThreads (deterministic).
-  // 2) Fallback to unique-candidate mapping only when there is exactly one safe target.
-  // Never guess by taking the first candidate.
-  useEffect(() => {
-    for (const task of kanbanTasks) {
-      if (!task.threadId) {
-        continue;
-      }
-      const canonicalThreadId = resolveTaskThreadId(task.threadId, resolveCanonicalThreadId);
-      if (canonicalThreadId && canonicalThreadId !== task.threadId) {
-        kanbanUpdateTask(task.id, { threadId: canonicalThreadId });
-        continue;
-      }
-
-      const taskThreadId = canonicalThreadId ?? task.threadId;
-      const isPendingThread =
-        taskThreadId.startsWith("claude-pending-") ||
-        taskThreadId.startsWith("opencode-pending-");
-      if (!isPendingThread) {
-        continue;
-      }
-      if (threadStatusById[taskThreadId] !== undefined) {
-        continue;
-      }
-      const wsId = workspacesByPath.get(task.workspaceId)?.id;
-      const threads = wsId ? (threadsByWorkspace[wsId] ?? []) : [];
-      if (threads.some((entry) => entry.id === taskThreadId)) {
-        continue;
-      }
-      const otherTaskThreadIds = new Set(
-        kanbanTasks
-          .filter((entry) => entry.id !== task.id && entry.threadId)
-          .map((entry) =>
-            resolveTaskThreadId(entry.threadId, resolveCanonicalThreadId),
-          )
-          .filter(
-            (threadId): threadId is string =>
-              Boolean(
-                threadId &&
-                  !threadId.startsWith("claude-pending-") &&
-                  !threadId.startsWith("opencode-pending-"),
-              ),
-          ),
-      );
-      const uniqueCandidate = resolvePendingSessionThreadCandidate({
-        pendingThreadId: taskThreadId,
-        workspaceThreadIds: threads.map((entry) => entry.id),
-        occupiedThreadIds: otherTaskThreadIds,
-      });
-      if (uniqueCandidate) {
-        kanbanUpdateTask(task.id, { threadId: uniqueCandidate });
-      }
-    }
-  }, [
-    kanbanTasks,
-    threadStatusById,
-    threadsByWorkspace,
-    kanbanUpdateTask,
-    workspacesByPath,
-    resolveCanonicalThreadId,
-  ]);
-
-  useEffect(() => {
-    if (appMode !== "kanban") {
-      setSelectedKanbanTaskId(null);
-    }
-  }, [appMode, setSelectedKanbanTaskId]);
-
-  // Sync activeWorkspaceId when kanban navigates to a workspace
-  useEffect(() => {
-    if (appMode === "kanban" && "workspaceId" in kanbanViewState) {
-      const kanbanWsPath = kanbanViewState.workspaceId;
-      const ws = kanbanWsPath ? workspacesByPath.get(kanbanWsPath) : null;
-      if (ws && ws.id !== activeWorkspaceId) {
-        setActiveWorkspaceId(ws.id);
-      }
-    }
-  }, [appMode, kanbanViewState, activeWorkspaceId, setActiveWorkspaceId, workspacesByPath]);
-
-  // Compute which kanban tasks are currently processing (AI responding)
-  const taskProcessingMap = useMemo(() => {
-    const map: Record<string, { isProcessing: boolean; startedAt: number | null }> = {};
-    for (const task of kanbanTasks) {
-      const taskThreadId = resolveTaskThreadId(task.threadId, resolveCanonicalThreadId);
-      if (taskThreadId) {
-        const status = threadStatusById[taskThreadId];
-        map[task.id] = {
-          isProcessing: status?.isProcessing ?? false,
-          startedAt: status?.processingStartedAt ?? null,
-        };
-      }
-    }
-    return map;
-  }, [kanbanTasks, threadStatusById, resolveCanonicalThreadId]);
-
-  useEffect(() => {
-    const runSchedulerTick = () => {
-      const nowTs = Date.now();
-      const activeTaskIds = new Set(kanbanTasksRef.current.map((entry) => entry.id));
-      for (const taskId of Object.keys(kanbanExecutionLocksRef.current)) {
-        if (activeTaskIds.has(taskId)) {
-          continue;
-        }
-        delete kanbanExecutionLocksRef.current[taskId];
-      }
-      for (const task of kanbanTasksRef.current) {
-        const runtimeLock = kanbanExecutionLocksRef.current[task.id];
-        if (runtimeLock) {
-          const hasPersistedExecutionLock = Boolean(task.execution?.lock);
-          const isLockExpired = nowTs - runtimeLock.acquiredAt > KANBAN_EXECUTION_LOCK_STALE_MS;
-          if (!hasPersistedExecutionLock || task.status !== "todo" || isLockExpired) {
-            delete kanbanExecutionLocksRef.current[task.id];
-            if (task.execution?.lock) {
-              updateTaskExecution(task.id, { lock: null });
-            }
-          }
-        }
-        if (task.execution?.blockedReason === "scheduled_trigger_blocked") {
-          updateTaskExecution(task.id, { blockedReason: null });
-        }
-        const schedule = task.schedule;
-        if (!schedule || schedule.mode === "manual") {
-          continue;
-        }
-        if (schedule.paused) {
-          continue;
-        }
-        const taskThreadId = resolveTaskThreadId(task.threadId, resolveCanonicalThreadId);
-        if (taskThreadId && task.threadId && taskThreadId !== task.threadId) {
-          kanbanUpdateTask(task.id, { threadId: taskThreadId });
-        }
-        const isTaskProcessing = taskThreadId
-          ? (threadStatusById[taskThreadId]?.isProcessing ?? false)
-          : false;
-        const shouldPromoteTestingToTodo =
-          schedule.mode === "recurring" &&
-          schedule.recurringExecutionMode !== "new_thread" &&
-          task.status === "testing" &&
-          !isTaskProcessing &&
-          typeof schedule.nextRunAt === "number" &&
-          schedule.nextRunAt <= nowTs;
-        const normalizedStatus = shouldPromoteTestingToTodo ? "todo" : task.status;
-        if (normalizedStatus !== task.status) {
-          kanbanUpdateTask(task.id, { status: normalizedStatus });
-        }
-        if (normalizedStatus !== "todo") {
-          continue;
-        }
-
-        const missedRunResult = applyMissedRunPolicy(
-          task,
-          schedulerStartedAtRef.current,
-          nowTs,
-        );
-        let effectiveSchedule = schedule;
-        if (missedRunResult) {
-          effectiveSchedule = missedRunResult.schedule;
-          kanbanUpdateTask(task.id, {
-            schedule: missedRunResult.schedule,
-          });
-          updateTaskExecution(task.id, {
-            lastSource: "scheduled",
-            blockedReason: missedRunResult.blockedReason,
-          });
-          continue;
-        }
-
-        if (!isScheduleDue(effectiveSchedule, nowTs)) {
-          continue;
-        }
-
-        if (
-          effectiveSchedule.mode === "recurring" &&
-          effectiveSchedule.recurringExecutionMode === "new_thread"
-        ) {
-          const recurringSeriesId =
-            typeof effectiveSchedule.seriesId === "string" && effectiveSchedule.seriesId.trim().length > 0
-              ? effectiveSchedule.seriesId.trim()
-              : task.id;
-          const hasSiblingExecuting = kanbanTasksRef.current.some((entry) => {
-            if (entry.id === task.id) {
-              return false;
-            }
-            const siblingSchedule = entry.schedule;
-            if (
-              !siblingSchedule ||
-              siblingSchedule.mode !== "recurring" ||
-              siblingSchedule.recurringExecutionMode !== "new_thread"
-            ) {
-              return false;
-            }
-            const siblingSeriesId =
-              typeof siblingSchedule.seriesId === "string" && siblingSchedule.seriesId.trim().length > 0
-                ? siblingSchedule.seriesId.trim()
-                : entry.id;
-            if (siblingSeriesId !== recurringSeriesId) {
-              return false;
-            }
-            return (
-              entry.status === "inprogress" ||
-              Boolean(kanbanExecutionLocksRef.current[entry.id])
-            );
-          });
-          if (hasSiblingExecuting) {
-            updateTaskExecution(task.id, {
-              lastSource: "scheduled",
-              blockedReason: "scheduled_trigger_blocked",
-            });
-            continue;
-          }
-        }
-
-        if (effectiveSchedule.mode === "recurring" && hasReachedRecurringRoundLimit(effectiveSchedule)) {
-          kanbanUpdateTask(task.id, {
-            status: "done",
-            schedule: {
-              ...effectiveSchedule,
-              nextRunAt: null,
-            },
-          });
-          updateTaskExecution(task.id, {
-            lastSource: "scheduled",
-            blockedReason: "max_rounds_reached_auto_completed",
-          });
-          continue;
-        }
-
-        if (isTaskProcessing || Boolean(kanbanExecutionLocksRef.current[task.id])) {
-          // Running/locked is an expected transient condition for due recurring tasks.
-          // Do not expose it as user-facing "blocked" state.
-          updateTaskExecution(task.id, { lastSource: "scheduled", blockedReason: null });
-          continue;
-        }
-
-        if (effectiveSchedule.mode === "once") {
-          const triggeredSchedule = markScheduleTriggered(
-            effectiveSchedule,
-            "scheduled",
-            nowTs,
-          );
-          kanbanUpdateTask(task.id, { schedule: triggeredSchedule });
-        } else {
-          kanbanUpdateTask(task.id, {
-            schedule: {
-              ...effectiveSchedule,
-              overdue: false,
-              lastTriggeredAt: nowTs,
-              lastTriggerSource: "scheduled",
-            },
-          });
-        }
-        updateTaskExecution(task.id, {
-          lastSource: "scheduled",
-          blockedReason: null,
-        });
-        const forceNewThread =
-          effectiveSchedule.mode === "recurring" &&
-          effectiveSchedule.recurringExecutionMode === "new_thread";
-        const injectedPrefix =
-          forceNewThread &&
-          effectiveSchedule.newThreadResultMode !== "none" &&
-          task.lastResultSnapshot
-            ? buildChainedPromptPrefix(task.lastResultSnapshot)
-            : undefined;
-        void launchKanbanTaskExecution({
-          taskId: task.id,
-          source: "scheduled",
-          activate: false,
-          forceNewThread,
-          injectedPrefix,
-        });
-      }
-    };
-
-    runSchedulerTick();
-    const timer = window.setInterval(runSchedulerTick, KANBAN_SCHEDULER_INTERVAL_MS);
-    return () => {
-      window.clearInterval(timer);
-    };
-  }, [
-    threadStatusById,
-    kanbanUpdateTask,
-    updateTaskExecution,
-    launchKanbanTaskExecution,
-    resolveCanonicalThreadId,
-    kanbanCreateTask,
-  ]);
-
-  // Track previous processing state to detect transitions
-  const prevProcessingMapRef = useRef<Record<string, boolean>>({});
-  const prevTaskStatusMapRef = useRef<Record<string, KanbanTaskStatus>>({});
-
-  useEffect(() => {
-    const previousStatusMap = prevTaskStatusMapRef.current;
-    const nextStatusMap: Record<string, KanbanTaskStatus> = {};
-    for (const task of kanbanTasks) {
-      nextStatusMap[task.id] = task.status;
-      const previousStatus = previousStatusMap[task.id];
-      if (previousStatus === task.status) {
-        continue;
-      }
-      if (task.status === "inprogress") {
-        const hasStartedAt = typeof task.execution?.startedAt === "number";
-        const hasFinishedAt = typeof task.execution?.finishedAt === "number";
-        if (!hasStartedAt || hasFinishedAt) {
-          updateTaskExecution(task.id, {
-            startedAt: Date.now(),
-            finishedAt: null,
-          });
-        }
-        continue;
-      }
-      if (previousStatus === "inprogress") {
-        updateTaskExecution(task.id, {
-          finishedAt: Date.now(),
-        });
-      }
-    }
-    prevTaskStatusMapRef.current = nextStatusMap;
-  }, [kanbanTasks, updateTaskExecution]);
-
-  useEffect(() => {
-    const prev = prevProcessingMapRef.current;
-    for (const task of kanbanTasks) {
-      const wasProcessing = prev[task.id] ?? false;
-      const nowProcessing = taskProcessingMap[task.id]?.isProcessing ?? false;
-      if (wasProcessing === nowProcessing) continue;
-
-      // AI finished processing (true → false): auto-move inprogress → testing
-      if (wasProcessing && !nowProcessing && task.status === "inprogress") {
-        const completedAt = Date.now();
-        updateTaskExecution(task.id, {
-          finishedAt: completedAt,
-        });
-        const nextStatus = resolvePostProcessingStatus(task);
-        if (task.schedule?.mode === "recurring") {
-          const completionSource = task.execution?.lastSource ?? "scheduled";
-          const recurringSignature = [
-            task.workspaceId,
-            task.panelId,
-            task.title,
-            String(task.schedule.interval ?? 1),
-            task.schedule.unit ?? "days",
-            task.schedule.newThreadResultMode ?? "pass",
-          ].join("|");
-          const recurringSiblings = kanbanTasksRef.current.filter((entry) => {
-            const schedule = entry.schedule;
-            if (
-              !schedule ||
-              schedule.mode !== "recurring" ||
-              schedule.recurringExecutionMode !== "new_thread"
-            ) {
-              return false;
-            }
-            const signature = [
-              entry.workspaceId,
-              entry.panelId,
-              entry.title,
-              String(schedule.interval ?? 1),
-              schedule.unit ?? "days",
-              schedule.newThreadResultMode ?? "pass",
-            ].join("|");
-            return signature === recurringSignature;
-          });
-          const siblingSeriesIds = Array.from(new Set(
-            recurringSiblings
-              .map((entry) => entry.schedule?.seriesId)
-              .filter((seriesId): seriesId is string =>
-                typeof seriesId === "string" && seriesId.trim().length > 0),
-          ));
-          const recurringSeriesId =
-            task.schedule.recurringExecutionMode === "new_thread"
-              ? (
-                task.schedule.seriesId ??
-                (siblingSeriesIds.length === 1 ? siblingSeriesIds[0] : null) ??
-                task.id
-              )
-              : task.schedule.seriesId ?? null;
-          if (task.schedule.recurringExecutionMode === "new_thread" && siblingSeriesIds.length <= 1) {
-            for (const sibling of recurringSiblings) {
-              if (!sibling.schedule || sibling.schedule.seriesId === recurringSeriesId) {
-                continue;
-              }
-              kanbanUpdateTask(sibling.id, {
-                schedule: {
-                  ...sibling.schedule,
-                  seriesId: recurringSeriesId,
-                },
-              });
-            }
-          }
-          const completedSchedule = markRecurringScheduleCompleted(
-            {
-              ...task.schedule,
-              seriesId: recurringSeriesId,
-            },
-            completionSource,
-            completedAt,
-          );
-          const reachedRoundLimit = hasReachedRecurringRoundLimit(completedSchedule);
-          if (task.schedule.recurringExecutionMode === "new_thread") {
-            kanbanUpdateTask(task.id, {
-              status: reachedRoundLimit ? "done" : nextStatus,
-              // Freeze this completed run card in review; next cycle will use a new cloned task.
-              schedule: {
-                ...completedSchedule,
-                nextRunAt: null,
-              },
-            });
-            if (!reachedRoundLimit) {
-              const hasPendingSeriesTask = recurringSiblings.some((sibling) => {
-                if (sibling.id === task.id) {
-                  return false;
-                }
-                const siblingSchedule = sibling.schedule;
-                if (
-                  !siblingSchedule ||
-                  siblingSchedule.mode !== "recurring" ||
-                  siblingSchedule.recurringExecutionMode !== "new_thread"
-                ) {
-                  return false;
-                }
-                const siblingSeriesId =
-                  typeof siblingSchedule.seriesId === "string" && siblingSchedule.seriesId.trim().length > 0
-                    ? siblingSchedule.seriesId.trim()
-                    : sibling.id;
-                if (siblingSeriesId !== recurringSeriesId) {
-                  return false;
-                }
-                return sibling.status === "todo" || sibling.status === "inprogress";
-              });
-              if (!hasPendingSeriesTask) {
-                kanbanCreateTask({
-                  workspaceId: task.workspaceId,
-                  panelId: task.panelId,
-                  title: task.title,
-                  description: task.description,
-                  engineType: task.engineType,
-                  modelId: task.modelId,
-                  branchName: task.branchName,
-                  images: task.images ?? [],
-                  autoStart: false,
-                  schedule: completedSchedule,
-                  chain: task.chain
-                    ? {
-                        ...task.chain,
-                        blockedReason: null,
-                      }
-                    : undefined,
-                });
-              }
-            } else {
-              updateTaskExecution(task.id, {
-                lastSource: completionSource,
-                blockedReason: "max_rounds_reached_auto_completed",
-              });
-            }
-          } else {
-            kanbanUpdateTask(task.id, {
-              status: reachedRoundLimit ? "done" : nextStatus,
-              schedule: reachedRoundLimit
-                ? {
-                    ...completedSchedule,
-                    nextRunAt: null,
-                  }
-                : completedSchedule,
-            });
-            if (reachedRoundLimit) {
-              updateTaskExecution(task.id, {
-                lastSource: completionSource,
-                blockedReason: "max_rounds_reached_auto_completed",
-              });
-            }
-          }
-        } else {
-          kanbanUpdateTask(task.id, { status: nextStatus });
-        }
-
-        const snapshot = extractKanbanResultSnapshot(
-          resolveTaskThreadId(task.threadId, resolveCanonicalThreadId),
-          (() => {
-            const taskThreadId = resolveTaskThreadId(task.threadId, resolveCanonicalThreadId);
-            return taskThreadId ? threadItemsByThread[taskThreadId] : undefined;
-          })(),
-        );
-        if (snapshot) {
-          const taskThreadId = resolveTaskThreadId(task.threadId, resolveCanonicalThreadId);
-          kanbanUpdateTask(task.id, {
-            ...(taskThreadId && task.threadId && taskThreadId !== task.threadId
-              ? { threadId: taskThreadId }
-              : null),
-            lastResultSnapshot: snapshot,
-          });
-        }
-
-        const downstreamTask = findTaskDownstream(kanbanTasksRef.current, task.id);
-        if (downstreamTask) {
-          if (!snapshot) {
-            setTaskChainBlockedReason(downstreamTask.id, "missing_upstream_snapshot");
-            updateTaskExecution(downstreamTask.id, {
-              lastSource: "chained",
-              blockedReason: "missing_upstream_snapshot",
-            });
-          } else if (downstreamTask.status !== "todo") {
-            setTaskChainBlockedReason(downstreamTask.id, "downstream_not_todo");
-            updateTaskExecution(downstreamTask.id, {
-              lastSource: "chained",
-              blockedReason: "downstream_not_todo",
-            });
-          } else if (downstreamTask.schedule?.mode && downstreamTask.schedule.mode !== "manual") {
-            setTaskChainBlockedReason(downstreamTask.id, "downstream_has_schedule");
-            updateTaskExecution(downstreamTask.id, {
-              lastSource: "chained",
-              blockedReason: "downstream_has_schedule",
-            });
-          } else {
-            setTaskChainBlockedReason(downstreamTask.id, null);
-            updateTaskExecution(downstreamTask.id, {
-              lastSource: "chained",
-              blockedReason: null,
-            });
-            void launchKanbanTaskExecution({
-              taskId: downstreamTask.id,
-              source: "chained",
-              activate: false,
-              injectedPrefix: buildChainedPromptPrefix(snapshot),
-            }).then((result) => {
-              if (!result.ok) {
-                setTaskChainBlockedReason(downstreamTask.id, result.reason);
-                updateTaskExecution(downstreamTask.id, {
-                  lastSource: "chained",
-                  blockedReason: result.reason,
-                });
-              }
-            });
-          }
-        }
-      }
-      // User sent follow-up (false → true): auto-move testing → inprogress
-      if (!wasProcessing && nowProcessing && task.status === "testing") {
-        kanbanUpdateTask(task.id, { status: "inprogress" });
-        updateTaskExecution(task.id, {
-          startedAt: taskProcessingMap[task.id]?.startedAt ?? Date.now(),
-          finishedAt: null,
-        });
-      }
-    }
-    const boolMap: Record<string, boolean> = {};
-    for (const [id, val] of Object.entries(taskProcessingMap)) {
-      boolMap[id] = val.isProcessing;
-    }
-    prevProcessingMapRef.current = boolMap;
-  }, [
+  const {
+    handleOpenTaskConversation,
+    handleRetryTaskRun,
+    handleResumeTaskRun,
+    handleCancelTaskRun,
+    handleForkTaskRun,
+    handleCloseTaskConversation,
+    handleKanbanCreateTask,
     taskProcessingMap,
-    kanbanTasks,
-    kanbanUpdateTask,
-    kanbanCreateTask,
-    threadItemsByThread,
-    setTaskChainBlockedReason,
-    updateTaskExecution,
-    launchKanbanTaskExecution,
-    resolveCanonicalThreadId,
-  ]);
-
-  useEffect(() => {
-    const store = loadTaskRunStore();
-    for (const run of store.runs) {
-      if (!run.linkedThreadId) {
-        continue;
-      }
-      if (run.status !== "planning" && run.status !== "running" && run.status !== "waiting_input") {
-        continue;
-      }
-      const canonicalThreadId = resolveTaskThreadId(run.linkedThreadId, resolveCanonicalThreadId);
-      if (!canonicalThreadId) {
-        continue;
-      }
-      const patch = deriveTaskRunTelemetryPatch({
-        run,
-        threadStatus: threadStatusById[canonicalThreadId],
-        items: threadItemsByThread[canonicalThreadId],
-        now: Date.now(),
-      });
-      if (!patch) {
-        continue;
-      }
-      const result = patchTaskRunAndProjectToKanban({
-        runId: run.runId,
-        status: patch.status,
-        currentStep: patch.currentStep,
-        latestOutputSummary: patch.latestOutputSummary,
-        artifacts: patch.artifacts,
-        finishedAt: patch.finishedAt,
-        now: patch.now,
-      });
-      if (result?.latestRunSummary) {
-        kanbanUpdateTask(run.task.taskId, {
-          latestRunSummary: result.latestRunSummary,
-        });
-      }
-    }
-  }, [
-    kanbanUpdateTask,
-    patchTaskRunAndProjectToKanban,
-    resolveCanonicalThreadId,
-    threadItemsByThread,
-    threadStatusById,
-  ]);
-
-  // Drag to "inprogress" auto-execute: create thread and send first message (without opening conversation panel)
-  const handleDragToInProgress = useCallback(
-    (task: KanbanTask) => {
-      void launchKanbanTaskExecution({
-        taskId: task.id,
-        source: "drag",
-        activate: false,
-      });
-    },
-    [launchKanbanTaskExecution],
-  );
+    handleDragToInProgress,
+  } = useAppShellKanbanExecutionSection(ctx);
 
   const orderValue = (entry: WorkspaceInfo) =>
     typeof entry.settings.sortOrder === "number"
