+++ src/app-shell-parts/useAppShellViewStateSection.ts
@@ -1,0 +1,97 @@
+import { useCallback, useEffect, useState } from "react";
+import { shouldHideHomeOnThreadActivation } from "../features/home/utils/homeVisibility";
+
+export function useAppShellViewStateSection({
+  activePlan,
+  activeTab,
+  activeThreadId,
+  activeWorkspace,
+  activeWorkspaceId,
+  appMode,
+  expandRightPanel,
+  homeOpen,
+  homeWorkspaceDefaultId,
+  isCompact,
+  isTablet,
+  selectedCollaborationMode,
+  setActiveThreadId,
+  setActiveWorkspaceId,
+  setHomeOpen,
+  tabletTab,
+}: any) {
+  const isPlanMode = selectedCollaborationMode?.mode === "plan";
+  const hasPlanData = Boolean(
+    activePlan && (activePlan.steps.length > 0 || activePlan.explanation)
+  );
+  const [isPlanPanelDismissed, setIsPlanPanelDismissed] = useState(false);
+  const hasActivePlan = hasPlanData && !isPlanPanelDismissed;
+  useEffect(() => {
+    setIsPlanPanelDismissed(false);
+  }, [activeThreadId]);
+  const openPlanPanel = useCallback(() => {
+    setIsPlanPanelDismissed(false);
+    expandRightPanel();
+  }, [expandRightPanel]);
+  const closePlanPanel = useCallback(() => {
+    setIsPlanPanelDismissed(true);
+  }, []);
+  const showKanban = appMode === "kanban";
+  const showGitHistory = appMode === "gitHistory";
+  const [selectedKanbanTaskId, setSelectedKanbanTaskId] = useState<string | null>(null);
+  const [workspaceHomeWorkspaceId, setWorkspaceHomeWorkspaceId] = useState<string | null>(null);
+  const showHome = (!activeWorkspace || homeOpen) && !showKanban;
+  const showWorkspaceHome = Boolean(
+    activeWorkspace &&
+      !showHome &&
+      workspaceHomeWorkspaceId === activeWorkspace.id &&
+      !activeThreadId &&
+      appMode === "chat" &&
+      (isCompact ? (isTablet ? tabletTab : activeTab) === "codex" : activeTab !== "spec"),
+  );
+  useEffect(() => {
+    if (!showHome || activeWorkspaceId || !homeWorkspaceDefaultId) {
+      return;
+    }
+    setActiveWorkspaceId(homeWorkspaceDefaultId);
+    setActiveThreadId(null, homeWorkspaceDefaultId);
+  }, [
+    activeWorkspaceId,
+    homeWorkspaceDefaultId,
+    setActiveThreadId,
+    setActiveWorkspaceId,
+    showHome,
+  ]);
+  useEffect(() => {
+    if (
+      !shouldHideHomeOnThreadActivation({
+        homeOpen,
+        activeThreadId,
+      })
+    ) {
+      return;
+    }
+    setHomeOpen(false);
+  }, [
+    activeThreadId,
+    homeOpen,
+    setHomeOpen,
+  ]);
+
+  return {
+    closePlanPanel,
+    hasActivePlan,
+    hasPlanData,
+    isPlanMode,
+    isPlanPanelDismissed,
+    openPlanPanel,
+    selectedKanbanTaskId,
+    setIsPlanPanelDismissed,
+    setSelectedKanbanTaskId,
+    setWorkspaceHomeWorkspaceId,
+    showGitHistory,
+    showHome,
+    showKanban,
+    showWorkspaceHome,
+    workspaceHomeWorkspaceId,
+  };
+}
